Move code from RecoverPasswordControllerComponent to DefaultRecoverPasswordControllerAction.
